Move TimeDiff component into its own module
